refactor(CubeAPI): extract exec-code service layer, sanitize rich output, centralise credentials
